Update psr/cache requirment to include v2 and v3
Also updates the cache implementation used in the example file
because tedvium/stash doesn't yet support the new versions of
psr/cacche.

Fix a bug with the cache example which was using the 'test'
API call, which declares itself as 'nocache'. Instead, call
it manually (with the default of nocache=false).

Refs #55